chore: extract skeleton also
