+++ apps/renderer/src/modules/entry-column/Items/index.ts
@@ -1,13 +1,13 @@
 import { FeedViewType } from "~/lib/enum"
 
-import { ArticleItem } from "./article-item"
-import { AudioItem } from "./audio-item"
-import { NotificationItem } from "./notification-item"
-import { PictureItem } from "./picture-item"
-import { SocialMediaItem } from "./social-media-item"
-import { VideoItem } from "./video-item"
+import { ArticleItem, ArticleItemSkeleton } from "./article-item"
+import { AudioItem, AudioItemSkeleton } from "./audio-item"
+import { NotificationItem, NotificationItemSkeleton } from "./notification-item"
+import { PictureItem, PictureItemSkeleton } from "./picture-item"
+import { SocialMediaItem, SocialMediaItemSkeleton } from "./social-media-item"
+import { VideoItem, VideoItemSkeleton } from "./video-item"
 
-const map = {
+const ItemMap = {
   [FeedViewType.Articles]: ArticleItem,
   [FeedViewType.SocialMedia]: SocialMediaItem,
   [FeedViewType.Pictures]: PictureItem,
@@ -16,5 +16,18 @@ const map = {
   [FeedViewType.Notifications]: NotificationItem,
 }
 export const getItemComponentByView = (view: FeedViewType) => {
-  return map[view] || ArticleItem
+  return ItemMap[view] || ArticleItem
+}
+
+const SkeletonItemMap = {
+  [FeedViewType.Articles]: ArticleItemSkeleton,
+  [FeedViewType.SocialMedia]: SocialMediaItemSkeleton,
+  [FeedViewType.Pictures]: PictureItemSkeleton,
+  [FeedViewType.Videos]: VideoItemSkeleton,
+  [FeedViewType.Audios]: AudioItemSkeleton,
+  [FeedViewType.Notifications]: NotificationItemSkeleton,
+}
+
+export const getSkeletonItemComponentByView = (view: FeedViewType) => {
+  return SkeletonItemMap[view]
 }
